Refactor code structure in algebra module
Reorganised the inclusion order of header files in multiple files in the algebra module for better readability and consistency. Cleaned up the format of the code by removing unnecessary white spaces, improving the argument layout in function declarations, and changing the way structs are styled for brevity and clarity. The changes made do not affect the functionality of the code.