Add detailed comments to intervals and streams classes
Documentation comments were added to several classes and methods in the 'intervals' and 'streams' namespaces. This includes in-depth descriptions and usage information for the 'Partition', 'RealInterval', and 'BrownianStream' classes, as well as methods such as 'refine_midpoints', 'mesh', 'insert_intermediate' and 'merge' in the 'Partition' class and the 'segment' function in the 'Segmentation' module.