Add host_kernel.h to roughpy device support
This update includes the addition of a new file named host_kernel.h to the roughpy device support. This new header file defines and implements a HostKernel class template, converted kernel argument classes, and related functions. It also includes the changes to the CMakeLists.txt file to properly build the project with this new addition.