Add debugging message for Python executable
This commit includes a check in the CMakeLists.txt file for the presence of a Python3 executable. If a Python3 executable is found, a debugging message will be printed that displays the path of the executable. This is useful for debugging purposes when trying to determine the path of the Python executable being used in the build process.