Refactor formatting and update KernelArgument class
The formatting in kernel_arg.h has been improved for readability. Additionally, the KernelArgument class has been updated. Information about the buffer type is stored in the argument object, even for Buffer arguments. This is useful for retrieving type information in subsequent stages, thus enhancing the efficiency of the entire process.