Build system changes (#99)
* change to auto-include vcpkg

* Disable JAX tests for interoperability issues

Interoperability with JAX is currently disabled, and tests depending on it are now skipped unconditionally. This ensures the test suite runs smoothly until the issue is resolved.

* Add Boost::url to streams/CMakeLists.txt

Boost::url is added to the list of public dependencies to ensure URL handling functionality is available. This change will allow the project to utilize Boost's URL library features effectively.