Fix GitHub actions again (#118)
* Update Dependabot auto-update schedule to monthly

Changed the auto-update interval for both pip and GitHub Actions dependencies from daily to monthly. This adjustment aims to reduce the frequency of dependency update checks and balance maintenance workload.

* Update macOS build matrix and artifact naming in CI

Revised the macOS versions to include Apple Silicon support and removed outdated OS tags. Also added a conditional step to install build dependencies on macOS and updated the artifact naming convention for better clarity.

* Update CI workflow to include additional paths and fix macOS step

Added more excluded paths to the CI workflow for efficiency. Corrected the case sensitivity issue for the macOS installation step to ensure proper execution on macOS runners.

* Install additional build dependencies for macOS

Added autoconf, automake, libtool, m4, and ninja to the build dependencies for macOS in the GitHub Actions workflow. This ensures necessary tools are available for building and testing on macOS runners.