Fix build time performance improvements (#161)
* Add precompiled headers for improved build performance

Introduce a new `RoughPy_PrecompiledHeaders` target to reduce build time, especially for cereal header parsing. Adjust CMakeLists for relevant modules to link against the new precompiled headers target.

* Update msbuild action in GitHub workflow to v2

Upgraded msbuild setup action to version 2 in the tests.yml file to ensure compatibility with the latest features and improvements. This change is scoped to the Windows runner environment in the GitHub CI workflow.