Fixed the degree bits in the libalgebra lite index_key type
The index_key is used for basis keys for libalgebra_lite tensors and Lie algebras. The default number of bits was set to 4, meaning a maximum of degree 15 for basis keys. This is probably the source of numerous errors in RoughPy when using degree >= 16. This change fixes the number of bits to 8, giving a maximum of degree 255, which should be completely sufficient until we can finally get rid of libalgebra_lite and replacing the badness within